FEAT: 비인기 여행지 피드 API 구현
유저가 등록한 태그로 검색한 경우, 해당 태그를 가진 비인기 여행지들을 나열

+++ spring-be/src/main/java/flaskspring/demo/tag/dto/req/ReqTagIndexes.java
@@ -10,5 +10,5 @@ import java.util.List;
 public class ReqTagIndexes {
 
     @Schema(description = "태그 인덱스 목록", example = "[1, 2, 3]")
-    private List<Integer> tagIndexes;
+    private List<Long> tagIndexes;
 }
